requirements.txt
Package                Version
---------------------- -------
cffi                   1.14.3
cryptography           3.2.1
mysql-connector-python 8.0.22
pip                    20.2.4
pkg-resources          0.0.0
protobuf               3.14.0
pycparser              2.20
PyMySQL                0.10.1
PyYAML                 5.3.1
setuptools             50.3.2
six                    1.15.0
